requirements.txt
Flask
flask_login
bcrypt
gunicorn
python-dotenv
openai